Update setup.sh
The source command is now using the install file located in the ros2 workspace that this setup.sh file is generating ("~/ros2_ws/install/setup.sh") instead of using the one located in the root folder ("/opt/ros/jazzy/setup.bash"). For some reason the original did not work in the system running ubuntu without VM software. If the problem occurs again, we might consider trying to solve this by adding a sourcing prompt for both of the files to .bashrc.